bug - silly programmer

+++ common/file_system_nuke/nuke_directory.ps1
@@ -2,8 +2,21 @@
 .EXAMPLE
   nuke_directory -dir_to_nuke:'F:\scm_databases\.git\' -names_to_leave:$null
 #>
-function nuke_directory ($dir_to_nuke, $names_to_leave)
+function nuke_directory
 {
+
+    [cmdletbinding()]
+    Param (
+        [Parameter(Mandatory=$true)][string]   $dir_to_nuke
+        ,[Parameter(Mandatory=$false)]object[]    $names_to_leave
+        ,[Parameter(Mandatory=$false)] [switch]    $leave_directory
+    )
+
+
+
+
+
+
     if ( $names_to_leave -eq $null ) { $names_to_leave = @() }
 
     $SCRIPT:files_in_use = $null
@@ -25,6 +38,11 @@ function nuke_directory ($dir_to_nuke, $names_to_leave)
           <# try simply deleting the directory #>       
           $delete_one_at_time = $false
         }
+
+        if ( $leave_directory -eq $true )
+        {
+          $delete_one_at_time = $true
+        }
   
         write-debug "     `$delete_one_at_time =[$delete_one_at_time]"
 
@@ -45,20 +63,23 @@ function nuke_directory ($dir_to_nuke, $names_to_leave)
         write-debug "     `$delete_one_at_time =[$delete_one_at_time]"
         if ( $delete_one_at_time -eq $true )
         {
-          $null = (nuke_directory_one_file_at_a_time -dir_to_nuke:$dir_to_nuke  -names_to_leave:$names_to_leave)
-          write-debug "     All Files are now probably gone.  Trying to delete the entire directory again."
-          if ($dir_to_remove_info.Exists -eq $true)
-            {
-              try 
+          if ( $leave_directory -eq $false )
+          {
+            $null = (nuke_directory_one_file_at_a_time -dir_to_nuke:$dir_to_nuke  -names_to_leave:$names_to_leave)
+            write-debug "     All Files are now probably gone.  Trying to delete the entire directory again."
+            if ($dir_to_remove_info.Exists -eq $true)
               {
-                $dir_to_remove_info.Delete($true) 
+                try 
+                {
+                  $dir_to_remove_info.Delete($true) 
+                }
+                catch {[Exception]}
+                {
+                  GET-Openfile -Filename:$dir_to_remove_info.FullName | Format-Table | Write-Host
+                  throw $_.Exception.Message
+                }
               }
-              catch {[Exception]}
-              {
-                GET-Openfile -Filename:$dir_to_remove_info.FullName | Format-Table | Write-Host
-                throw $_.Exception.Message
-              }
-            }
+          }          
         }
     }
     else
